integrate cli_test.go
